Services and Contact section moved in common folder
