+++ src/components/common/services/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Grid, Typography, Box } from '@mui/material';
 import { makeStyles } from '@mui/styles';
-import ServicesBox from 'src/components/home/services/services';
+import ServicesBox from 'src/components/common/services/services';
 import { WebDevIcon } from 'src/components/svg/webDevIcon';
 import servicesData from 'src/components/home/services/services.json';
